chore(artifacts): make 404 errors show their backend error for clarity
