+++ wandb/apis/normalize.py
@@ -3,32 +3,16 @@
 import ast
 import sys
 from functools import wraps
-from typing import Callable, List, TypeVar
+from typing import Callable, TypeVar
 
 import requests
 from wandb_gql.client import RetryError
 
 from wandb import env
 from wandb.errors import CommError, Error
+from wandb.util import parse_backend_error_messages
 
 _F = TypeVar("_F", bound=Callable)
-
-
-def parse_backend_error_messages(response: requests.Response) -> List[str]:
-    errors = []
-    try:
-        data = response.json()
-    except ValueError:
-        return errors
-
-    if "errors" in data and isinstance(data["errors"], list):
-        for error in data["errors"]:
-            # Our tests and potentially some api endpoints return a string error?
-            if isinstance(error, str):
-                error = {"message": error}
-            if "message" in error:
-                errors.append(error["message"])
-    return errors
 
 
 def normalize_exceptions(func: _F) -> _F:
